fixed removal of the .html file in overwrite mode.
git-svn-id: http://svn.cern.ch/guest/lhcb/LbScripts/trunk@120458 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 3732efa9d88faa2c72c30403ae1d8aff3a7eb8c2

+++ LbScripts/LbLegacy/python/LbLegacy/install_project.py
@@ -857,7 +857,7 @@ def getProjectList(name, version, binary=None, recursive=True):
 
     if not check_only :
         checkWriteAccess(this_html_dir)
-        if overwrite_mode :
+        if overwrite_mode and os.path.exists(tar_file_html) :
             os.remove(tar_file_html)
         getFile(url_dist + "html/", tar_file_html)
